With multiple clinics/clearinghouses and one virtual database, reports no longer get pulled in automatically. Deleting a supplier or supply item is now checked first for dependent keys.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@1179 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Data Interface/Suppliers.cs
@@ -25,7 +25,16 @@ namespace OpenDental{
 		///<summary>Surround with try-catch.</summary>
 		public static void DeleteObject(Supplier supp){
 			//validate that not already in use.
-
+			string command="SELECT COUNT(*) FROM supplyorder WHERE SupplierNum="+POut.PInt(supp.SupplierNum);
+			int count=PIn.PInt(General.GetCount(command));
+			if(count>0) {
+				throw new ApplicationException(Lan.g("Supplies","Supplier is already in use on an order. Not allowed to delete."));
+			}
+			command="SELECT COUNT(*) FROM supply WHERE SupplierNum="+POut.PInt(supp.SupplierNum);
+			count=PIn.PInt(General.GetCount(command));
+			if(count>0) {
+				throw new ApplicationException(Lan.g("Supplies","Supplier is already in use on a supply. Not allowed to delete."));
+			}
 			DataObjectFactory<Supplier>.DeleteObject(supp);
 		}
